move contactinfo into cards component
