+++ src/components/Card/index.jsx
@@ -46,4 +46,18 @@ const ProjectCard = ({ image, title, link, github }) => {
   );
 };
 
-export { Card, ProjectCard };
+const ContactInfo = ({ icon, title, link, linkText }) => {
+  return (
+    <div className={styles.infoDiv}>
+      {icon}
+      <div className={styles.info}>
+        <span>{title}</span>
+        <a href={link} target="_blank" rel="noreferrer">
+          {linkText}
+        </a>
+      </div>
+    </div>
+  );
+};
+
+export { Card, ProjectCard, ContactInfo };
