--- src/components/ContactInfo/ContactInfo.module.css
@@ -1,49 +1,0 @@
-.infoDiv {
-  margin-bottom: 1rem;
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  width: fit-content;
-}
-
-.infoDiv:hover {
-  transform: scale(1.1);
-}
-
-.infoDiv:hover > svg {
-  color: blueviolet;
-}
-
-.info {
-  display: flex;
-  flex-direction: column;
-}
-
-.info > span {
-  font-size: 1.3rem;
-  font-weight: 600;
-}
-
-.info > a {
-  text-decoration: none;
-  color: var(--text-color);
-  font-weight: 400;
-  opacity: 0.8;
-
-  transition: color 0.2s ease-in-out;
-}
-
-.info > a:hover {
-  text-decoration: underline;
-  color: blueviolet;
-}
-
-@media only screen and (max-width: 768px) {
-  .info > span {
-    font-size: 1.1rem;
-  }
-
-  .info > a {
-    font-size: 0.9rem;
-  }
-}
